requirements.txt
gradio==5.42.0
ollama==0.5.3
pandas==2.3.1
python-dotenv==1.1.1
requests==2.32.4
